Start of Account Access
Started creating dialog boxes for the account page. Have a dialog box
for: signing in, creating an account, saying the user logged in, and
saying the user failed at logging in. Created a whole new page that the
user has access to once they have logged in. Started creating the
framework that will be used to store the user information and shopping
cart information. Added a simple logout button on the account page.